Move apiFetch middleware into a separate package and add a disable method
